# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Devices API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/user/devices/{userDeviceID}/integrations/{integrationID}/commands/burn": {
            "get": {
                "description": "Produces the payload that the user signs and submits to burn a synthetic device.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user device KSUID, must be minted",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "integration KSUD, must be software-based and active",
                        "name": "integrationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apitypes.TypedData"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Submit the signature required for the synthetic device burning meta-transaction.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user device KSUID, must be minted",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "integration KSUD, must be software-based and active",
                        "name": "integrationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "only field is the signed EIP-712",
                        "name": "signed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.BurnSyntheticDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apitypes.TypedData"
                            }
                        }
                    }
                }
            }
        },
        "/user/devices/{userDeviceID}/integrations/{integrationID}/commands/mint": {
            "get": {
                "description": "Produces the payload that the user signs and submits to mint a synthetic device for\nthe given vehicle and integration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user device KSUID",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "integration KSUD, must be software-based",
                        "name": "integrationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apitypes.TypedData"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Submit a metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user device KSUID, must be minted",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "integration KSUD, must be software-based",
                        "name": "integrationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "only field is the signed EIP-712",
                        "name": "signed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.MintSyntheticDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apitypes.Type": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "apitypes.TypedData": {
            "type": "object",
            "properties": {
                "domain": {
                    "$ref": "#/definitions/apitypes.TypedDataDomain"
                },
                "message": {
                    "$ref": "#/definitions/apitypes.TypedDataMessage"
                },
                "primaryType": {
                    "type": "string"
                },
                "types": {
                    "$ref": "#/definitions/apitypes.Types"
                }
            }
        },
        "apitypes.TypedDataDomain": {
            "type": "object",
            "properties": {
                "chainId": {
                    "$ref": "#/definitions/math.HexOrDecimal256"
                },
                "name": {
                    "type": "string"
                },
                "salt": {
                    "type": "string"
                },
                "verifyingContract": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "apitypes.TypedDataMessage": {
            "type": "object",
            "additionalProperties": true
        },
        "apitypes.Types": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "$ref": "#/definitions/apitypes.Type"
                }
            }
        },
...
        "internal_controllers.BurnSyntheticDeviceRequest": {
            "type": "object",
            "properties": {
                "signature": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.MintSyntheticDeviceRequest": {
            "type": "object",
            "properties": {
                "signature": {
                    "type": "string",
                    "example": "0xc565d38982e1a5004efb5ee390fba0a08bb5e72b3f3e91094c66bc395c324f785425d58d5c1a601372d9c16164e380c63e89f1e0ea95fdefdf7b2854c4f938e81b"
                }
            }
        },
...
        "math.HexOrDecimal256": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}